ref: move DIG parser to a helper file
